Fix height issues when using a TokenEdit as an item delegate
I believe the issue is that `setFixedHeight` sets both the minimum and
maximum heights, but the conditional only checks if the widget's actual
height is different than the new height, then some layout logic resizes
the TokenEdit to match its maximum height, making it much larger than
it's supposed to be. But that's just a guess. In any case, this fixes
the problem and skipping the height check seems pretty harmless.